BREAKING Added proper defer feature
BREAKING CHANGE

The this.ready and this.emit("ready") were tests I was making to try and figure out how to wait for the collection to have loaded all data before proceeding. I have now figured that out. 

The change removes both those properties in favour of a simple promise that is resolved after init. Feel free to steal that trick by the way, it's the `this.defer` line and `this.ready()` in the init! 

This means you can just wait for the promise to resolve before using the PCollection: 
```
const myColl = new PersistentCollection({name: "test"});
myColl.defer.then( () => {
  // collection is full and ready, whatever amount of items it has
});
```
If in an async method (yay node 7.6+), just use `await myColl.defer;`